# internal/models/subscription.go
package models

import (
	"time"

	"github.com/google/uuid"
)

type Subscription struct {
	ID          int       `json:"id" gorm:"primaryKey"`
	ServiceName string    `json:"service_name" binding:"required"`
	Price       int       `json:"price" binding:"required,min=0"`
	UserID      uuid.UUID `json:"user_id" gorm:"type:uuid" binding:"required"`
	StartDate   string    `json:"start_date" binding:"required"`
	EndDate     *string   `json:"end_date,omitempty" `
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
}